requirements.txt
apache-airflow-providers-airbyte==5.4.0
yfinance==1.2.1
psycopg2-binary==2.9.10